requirements.txt
numpy==2.1.3
opencv-python==4.10.0.84
pyyaml==6.0.2
tqdm==4.66.5
rich==13.9.4
torch
ultralytics
deep-sort-realtime
